Separate merge from handle
* AbstractKirinModelBuilder provides merge_trip_updates()
  simply calling merge() fn for now
* handler.merge() moved in merge_utils to be reused by KirinModelBuilders
* create build_wrapper to host fn using KirinModelBuilder
  (move wrap_build & handle in it)
