+++ kirin/core/__init__.py
@@ -30,4 +30,3 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
-from kirin.core.handler import handle
