+++ kirin/core/abstract_builder.py
@@ -31,79 +31,16 @@
 
 from __future__ import absolute_import, print_function, unicode_literals, division
 
-import logging
 from abc import ABCMeta
-from datetime import datetime
 from typing import Tuple, List, Dict, Any
 
 import navitia_wrapper
 import six
 from flask import current_app
 
-from kirin import core, redis_client
-from kirin.core import model
+from kirin import redis_client
+from kirin.core.merge_utils import merge
 from kirin.core.model import Contributor, RealTimeUpdate, TripUpdate
-from kirin.exceptions import KirinException
-from kirin.new_relic import is_invalid_input_exception, record_custom_parameter
-from kirin.utils import set_rtu_status_ko, allow_reprocess_same_data, record_call
-
-
-def wrap_build(builder, input_raw):
-    """
-    Function wrapping the processing of realtime information of an external feed
-    This manages errors/logger/newrelic
-    :param builder: the KirinModelBuilder to be called
-    :param input_raw: the feed to process
-    """
-    contributor = builder.contributor
-    start_datetime = datetime.utcnow()
-    rt_update = None
-    log_dict = {"contributor": contributor.id}
-    status = "OK"
-
-    try:
-        # create a raw rt_update obj, save the raw_input into the db
-        rt_update, rtu_log_dict = builder.build_rt_update(input_raw)
-        log_dict.update(rtu_log_dict)
-
-        # raw_input is interpreted
-        trip_updates, tu_log_dict = builder.build_trip_updates(rt_update)
-        log_dict.update(tu_log_dict)
-
-        # finally confront to previously existing information (base_schedule, previous real-time)
-        _, handler_log_dict = core.handle(rt_update, trip_updates, contributor.id, builder.is_new_complete)
-        log_dict.update(handler_log_dict)
-
-    except Exception as e:
-        status = "failure"
-        allow_reprocess = True
-        if is_invalid_input_exception(e):
-            status = "warning"  # Kirin did his job correctly if the input is invalid and rejected
-            allow_reprocess = False  # reprocess is useless if input is invalid
-
-        if rt_update is not None:
-            error = e.data["error"] if (isinstance(e, KirinException) and "error" in e.data) else e.message
-            set_rtu_status_ko(rt_update, error, is_reprocess_same_data_allowed=allow_reprocess)
-            model.db.session.add(rt_update)
-            model.db.session.commit()
-        else:
-            # rt_update is not built, make sure reprocess is allowed
-            allow_reprocess_same_data(contributor.id)
-
-        log_dict.update({"exc_summary": six.text_type(e), "reason": e})
-
-        record_custom_parameter("reason", e)  # using __str__() here to have complete details
-        raise  # filters later for APM (auto.)
-
-    finally:
-        log_dict.update({"duration": (datetime.utcnow() - start_datetime).total_seconds()})
-        record_call(status, **log_dict)
-        if status == "OK":
-            logging.getLogger(__name__).info(status, extra=log_dict)
-        elif status == "warning":
-            logging.getLogger(__name__).warning(status, extra=log_dict)
-        else:
-            logging.getLogger(__name__).error(status, extra=log_dict)
 
 
 class AbstractKirinModelBuilder(six.with_metaclass(ABCMeta, object)):
@@ -146,3 +83,15 @@ class AbstractKirinModelBuilder(six.with_metaclass(ABCMeta, object)):
         :return log_dict: dict of (k,v) to be displayed in logs and newrelic
         """
         raise NotImplementedError("Please implement this method")
+
+    def merge_trip_updates(self, navitia_vj, db_trip_update, new_trip_update):
+        # type: (Dict[unicode, Any], TripUpdate, TripUpdate) -> TripUpdate
+        """
+        Merges information from:
+        * navitia_vj (base-schedule VJ - complete, maybe nonexistent)
+        * db_trip_update (last known realtime VJ - complete, maybe nonexistent)
+        * new_trip_update (VJ information extracted from incoming feed - maybe incomplete)
+        Returns resulting TripUpdate:
+        usually update of the last known realtime VJ, or completed version of new_trip_update
+        """
+        return merge(navitia_vj, db_trip_update, new_trip_update, self.is_new_complete)
