+++ tests/integration/gtfs_rt_test.py
@@ -36,7 +36,7 @@ import datetime
 import pytest
 
 from kirin.core import model
-from kirin.core.abstract_builder import wrap_build
+from kirin.core.build_wrapper import wrap_build
 from kirin.core.model import (
     RealTimeUpdate,
     db,
